Fix ESLint configuration for ES module support
Update ESLint configuration to use ES module syntax and resolve import statement issue.

* Delete `eslint.config.js` file.
* Modify `.eslintrc.js` file:
  - Add `import` statements for `FlatCompat`, `fileURLToPath`, `path`, `@typescript-eslint/eslint-plugin`, `@typescript-eslint/parser`, and `eslint-plugin-prettier`.
  - Initialize `compat` variable using `FlatCompat` with `baseDirectory` and `resolvePluginsRelativeTo` options.
  - Replace `module.exports` object with `compat.config` array containing the ESLint configuration.
  - Update ESLint configuration to use ES module syntax and include necessary plugins and rules.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Bacze12/inventory-pos?shareId=XXXX-XXXX-XXXX-XXXX).
